Memory optimisation: reduce peak RAM by ~500-700 MB per image
Based on a systematic audit of peak memory usage across the pipeline.

## main.py (~300 MB saved per image)
- image_for_limits: replace np.where(mask, nan, img).astype(float32) with
  float32 copy + in-place masking. The previous pattern created a full float64
  temporary array (~128 MB) before the astype conversion.
- image_orig / result_orig: add del immediately after PSF build from original
  image so the ~64 MB science frame is freed before aligned-image processing.
- Weight maps (sci + ref): dtype=float -> float32, saving ~64 MB each.
- Target cutout + RMS cutout: dtype=float -> float32 (~1-2 MB each, but called
  per-target across many images).
- LPI stamp copies: dtype=float -> float32 (called repeatedly in LPI loop).
- Gaussian-fit cutout: dtype=float -> float32.

## background.py (~320 MB saved per image)
- Iterative-masking residual array: float64 -> float32 (~64 MB for 4k×4k image;
  float32 is fully sufficient for sigma-clip source detection).
- bkg_data extraction from Background2D: dtype=float -> float32 (~64-128 MB).
- Flat-fallback bkg_surface + bkg_rms: dtype=float -> float32 (~128 MB combined).

## psf.py
- MCMC chain: add del chain, logp after summary stats are extracted. The raw
  chain (50 walkers × 5000 steps × 4 params ≈ 8 MB) accumulated across all
  sources in a batch — now freed immediately after percentiles are computed.
- Projection plot: convert data to float once, reuse for both axes, then del.
  Previously two asarray(dtype=float) calls created two full copies of the PSF.

## zeropoint.py
- Robust MCMC: add del samples, zp_samples, f_out_samples after extraction.
- Standard MCMC: remove redundant .flatten() (flat=True already returns 1-D);
  add del samples after statistics are computed. Each chain is ~4-8 MB; across
  N images this prevents unbounded accumulation.

+++ background.py
@@ -434,12 +434,14 @@ class BackgroundSubtractor:
         )
 
         mask = np.zeros(image.shape, dtype=bool)
-        # Avoid an unconditional full copy when possible (NumPy 2.0+ supports copy=).
+        # float32 is sufficient precision for iterative source-masking.
+        # Using float32 halves the memory footprint vs float64 (~64 MB saved
+        # for a 4096×4096 image).  Avoid an unconditional full copy when possible.
         try:
-            residual = np.asarray(image, dtype=np.float64, copy=False)
+            residual = np.asarray(image, dtype=np.float32, copy=False)
         except (TypeError, ValueError):
             # NumPy <2.0 raises TypeError for copy= keyword; NumPy 2.0+ raises ValueError if copy cannot be avoided
-            residual = np.asarray(image, dtype=np.float64)
+            residual = np.asarray(image, dtype=np.float32)
 
         # Safety: cap mask fraction so Background2D always has enough data to
         # fit a gradient.  If the mask exceeds this, stop iterating.
@@ -1117,7 +1119,7 @@ class BackgroundSubtractor:
                 # Use the raw Background2D background map without any
                 # additional median or Gaussian post-processing so that
                 # execution time stays minimal.
-                bkg_data = np.asarray(bkg.background, dtype=float)
+                bkg_data = np.asarray(bkg.background, dtype=np.float32)
 
                 self.logger.info(border_msg("Background estimation"))
                 self.logger.info(
@@ -1150,8 +1152,8 @@ class BackgroundSubtractor:
         # All attempts failed - flat fallback.
         self.logger.warning("All Background2D attempts failed - using global stats")
         gmean, gmed, gstd = sigma_clipped_stats(image, sigma=3.0, mask=mask)
-        bkg_surface = np.full_like(image, gmed, dtype=float)
-        bkg_rms = np.full_like(image, max(gstd, 1.0), dtype=float)
+        bkg_surface = np.full_like(image, gmed, dtype=np.float32)
+        bkg_rms = np.full_like(image, max(gstd, 1.0), dtype=np.float32)
         return True, bkg_surface, bkg_rms, gmed
 
     # -------------------------------------------------------------------------
